Add files via upload
fixes #1 -- Improved initialization system. The old system attached to the window onload event which works well but delays the start of script loading until the Dom AND all resources were finished. The new system attaches to tthe domContentLoaded event if available or the readyStateChange event as a backup. Both of these allow the loader to start as soon as the DOM is ready without waiting for resources. However, in the case that neither of these new methods are available, the original system remains as a fallback option.