Add files via upload
Improved init system. The old system had the code for attaching the init event to dom load was written in-line in the initialization section of the class. This functionality was moved to a function that is called in the initialization section. This makes the code cleaner and easier to read.