Add files via upload
Improved init system. Previously, very little validation was done to the config object that is passed as an argument to the constructor. If any values were invalid, they were simply discarded and a default used. Some simple validations have been added and the new system is set to throw errors on invalid values once the error system is complete. Additional validations may be necessary to ensure proper usage of the loader.